Refs #1038, add the referendum to the public proposal's timeline, and add the public proposal to the referendum's timeline
